Refactor locals processing, move hcl utilities to own package
* move locals processing to its own package
* simplify it by losing reserved word processing, since now we do not allow shadowing anyway
* dispense with the source finder since not realy needed, given full range in the diags
* extract commit hcl stuff into an hclutils package
